Add Base64 encode/decode support
Prior to this commit the cryptogram may be returned as binary data
that yaml will not be able to parse. This adds a pre and post action
Base64. This is based on design guidelines , rather then data yet. So
will need to see if this is actually needed ( for all mechanisms ?)