(PE-9993) Add `mri-pkcs11` mode
This adds a new `mode` called `mri-pkcs11`.  This mode is intended
to be used when running under Puppet Server, where we are not able
to use the native pkcs11 gem to handle the encryption.  When using
this mode, we'll use some JRuby internals to spawn an external
process, where we'll call MRI ruby and have it do the encryption/decryption.